+++ Hearthstone Deck Tracker/LogReader/Interfaces/IHsGameState.cs
@@ -25,6 +25,7 @@ namespace Hearthstone_Deck_Tracker.LogReader.Interfaces
 		int JoustReveals { get; set; }
 		Dictionary<int, IList<(string, DeckLocation, string?, EntityInfo?)>> KnownCardIds { get; set; }
 		int LastCardPlayed { get; set; }
+		int LastEntityChosenOnDiscover { get; set; }
 		Stack<string> LastPlagueDrawn { get; set; }
 		bool WasInProgress { get; set; }
 		int GameTriggerCount { get; set; }